+++ packages/danger_dart/lib/commands/process_command.dart
@@ -15,10 +15,8 @@ class ProcessCommand extends Command {
   final DangerUtil _dangerUtil;
   final Stdin _stdin;
   final Stdout _stdout;
-  final bool shouldExitOnEnd;
 
-  ProcessCommand(this._dangerUtil, this._stdin, this._stdout,
-      {this.shouldExitOnEnd = true}) {
+  ProcessCommand(this._dangerUtil, this._stdin, this._stdout) {
     argParser.addOption(
       'dangerfile',
       help: 'Location of dangerfile',
@@ -40,7 +38,8 @@ class ProcessCommand extends Command {
   @override
   Future<void> run() async {
     final args = argResults;
-    final str = (await _stdin.transform(utf8.decoder).toList()).join('');
+
+    var inputStr = (await _stdin.transform(utf8.decoder).toList()).join('');
 
     final isVerbose = args.wasParsed('verbose');
     final useColors = (Platform.environment['TERM'] ?? '').contains('xterm');
@@ -51,14 +50,34 @@ class ProcessCommand extends Command {
           DebugTree(useColors: useColors, logLevels: ['I', 'W', 'E']));
     }
 
-    if (str.isEmpty) {
-      throw 'Data not found';
+    if (inputStr.isEmpty) {
+      _stdout.write('danger://send-dsl');
+      inputStr = (await _stdin.transform(utf8.decoder).toList()).join('');
     }
+
+    if (inputStr.isEmpty) {
+      throw 'STDIN Error get empty input';
+    }
+
+    if(!inputStr.startsWith(r'danger://dsl/')) {
+      throw 'STDIN Error expect input [danger://dsl/] but got [$inputStr]';
+    }
+
+    final inputFilePath = inputStr.substring('danger://dsl/'.length);
+    final inputFile = File(inputFilePath);
+
+    if(!inputFile.existsSync()) {
+      throw 'File not found [$inputFilePath]';
+    }
+
+    final str = inputFile.readAsStringSync();
 
     final dangerFile = args['dangerfile'] as String;
     if (dangerFile == null) {
       throw 'Dangerfile not found';
     }
+
+    DangerIsolateReceiver isolateReceiver;
 
     try {
       final json = jsonDecode(str);
@@ -66,23 +85,29 @@ class ProcessCommand extends Command {
       final _ = DangerJSON.fromJson(json);
 
       final filePath = Uri.parse(join(current, dangerFile));
-      final isolateReceiver = DangerIsolateReceiver(json);
+      isolateReceiver = DangerIsolateReceiver(json);
 
       await _dangerUtil.spawnUri(filePath, isolateReceiver.toMessage());
 
-      _stdout.write(jsonEncode(isolateReceiver.dangerResults));
-      if (shouldExitOnEnd) {
-        exit(0);
+      final resultStr = jsonEncode(isolateReceiver.dangerResults);
+      final tempDir = Directory.systemTemp;
+      final tempFile = File(join(tempDir.path, 'danger-results.json'));
+      if(tempFile.existsSync()){
+        tempFile.deleteSync();
       }
+      tempFile.createSync();
+      tempFile.writeAsStringSync(resultStr);
+      
+      _stdout.write('danger-results:/${tempFile.path}');
+      await _stdout.flush();
     } catch (e) {
       if (e is Error) {
         _logger.e(e.toString(), ex: e, stacktrace: e.stackTrace);
       } else {
         _logger.e(e.toString(), ex: e);
       }
-      if (shouldExitOnEnd) {
-        exit(1);
-      }
+    } finally {
+      isolateReceiver?.receivePort?.close();
     }
   }
 }