+++ packages/danger_dart/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.5
+
+- Use passURLForDSL for local command
+
+## 0.0.4
+
+- Use passURLForDSL instead
+
+## 0.0.3
+
+- Close receive port before exit
+
+## 0.0.2
+
+- Fix Danger Process finished before writing all results
+
 ## 0.0.1
 
 - Initial version, created by Stagehand